static/static/requirements.txt
asgiref==3.4.1
awsebcli==3.20.3
backports.entry-points-selectable==1.1.0
beautiful-date==2.0.0
boto==2.49.0
boto3
botocore
cachetools==4.2.2
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.2
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.5
django-background-tasks==1.2.5
django-braces==1.14.0
django-ckeditor==6.1.0
django-compat==1.0.15
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-elasticsearch-dsl==7.2.1
django-elasticsearch-dsl-drf==0.22.2
django-filter==2.4.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-nine==0.2.5
django-oauth-toolkit==1.5.0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-allauth==0.45.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
elasticsearch==7.10.0
elasticsearch-dsl==7.4.0
filelock==3.0.12
future==0.16.0
gcsa==1.2.0
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.33.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
pathspec==0.9.0
phonenumbers==8.12.27
Pillow==8.3.1
platformdirs==2.0.2
protobuf==3.17.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyngrok==5.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
requests==2.26.0
requests-aws4auth==1.1.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
semantic-version==2.8.5
six==1.14.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.1
termcolor==1.1.0
twilio==7.4.0
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.9
virtualenv==20.6.0
wcwidth==0.1.9
whitenoise==5.3.0
wrapt==1.12.1
